Improve ESLint setup
I updated the config and tightened the ruleset. To the best of my
knowledge, this version represents what we're aiming for: we use
@typescript-eslint's plugin and parser (so that ESLint can process
TypeScript files) and extend several good rulesets.

I turned to "warn" most of the rules that are leftover because I'd like
to discuss them and address them in a future PR.