contracts: Rework public interface to act on pools directly
Beneficiary pool operations like staking, unstaking and claiming are now
called directly on the pool contracts, which also emit the events. This
means that the pool address for a particular beneficiary has to be known,
before, either by querying directly on the main contract, or by other
means, e.g. from a subgraph which tracks PoolDeployed events.

The main contract's aave pool variable got renamed to aavePool to avoid
confusion.