contracts: Add ToucanOffsetter implementation
The Toucan OffsetHelper contract (https://github.com/ToucanProtocol/OffsetHelper),
which is used in the ToucanOffsetter, was missing a method for swapping
an exact amount of input token.
I opened a PR to add this functionality at https://github.com/ToucanProtocol/OffsetHelper/pull/4
but at the time of this commit, this PR hasn't been merged yet. So the
ToucanOffsetter doesn't work on Polygon (Mumbai) yet.